Add export interventions api

+++ app/Infrastructure/Models/Amende.php
@@ -8,6 +8,6 @@ class Amende extends Model
 {
     protected $fillable = ['ordre', 'montant', 'compte_id', 'ecriture_categorie_id'];
     protected $casts = [
-        'ordre' => 'integer', 'montant' => 'float', 'compte_id' => 'integer', 'ecriture_categorie_id' => 'integer'
+        'ordre' => 'integer', 'montant' => 'numeric', 'compte_id' => 'integer', 'ecriture_categorie_id' => 'integer'
     ];
 }
